Refactor testPull structure
