+++ packages/data-broker/src/tests/services/dhis/DhisService/testPull/index.js
@@ -1,0 +1,1 @@
+export { testPull } from './testPull';
